+++ src/ionic-tree-provider.ts
@@ -133,8 +133,9 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
   private getFolderInfo(folder: string): FolderInfo {
     const packageJsonPath = path.join(this.workspaceRoot, 'package.json');
     const folders = isFolderBasedMonoRepo(this.workspaceRoot);
-    const folderBased = folders.length > 0;
     const packageJsonExists = this.pathExists(packageJsonPath);
+    const folderBased = folders.length > 0 && !packageJsonExists;
+
     return {
       packageJsonExists,
       folderBased,